requirements.txt
starlette
sse-starlette
requests
sseclient-py
